Front end fixed

+++ Moongy.RD.Launchpad.CodeGenerator.Tokenomics/Models/Tax/TaxTokenomicModel.cs
@@ -1,7 +1,7 @@
 ﻿namespace Moongy.RD.Launchpad.CodeGenerator.Tokenomics.Models.Tax;
 public class TaxTokenomicModel
 {
-    public int TaxFee { get; set; }
+    public double TaxFee { get; set; }
 
     public List<TaxRecipient> TaxRecipients { get; set; } = [];
 }
